SQF_bot/requirements.txt
ib_insync==0.9.86     # Broker API wrapper
pandas==2.2.2
numpy==1.26.4
apscheduler==3.10.4
pyyaml==6.0.1

